+++ src/TallTileStackDrawable.cpp
@@ -25,7 +25,7 @@ TallTileStackDrawable::TallTileStackDrawable(const TileStack& tileStack, const T
         if(tile->view().isTall())
         {
             m_indexOfFirstTallTile = i;
-            m_boundingRectangle = tile->model().collider() + Vec2F(tileLocation.x, tileLocation.y) * GameConstants::tileSize;
+            m_boundingRectangle = tile->model().collider() + Vec2F(static_cast<float>(tileLocation.x), static_cast<float>(tileLocation.y)) * static_cast<float>(GameConstants::tileSize);
             m_center = m_boundingRectangle.centerOfMass();
 
             break;
